wisper_bundle (Unzipped Files)/requirements.txt
# Only non-stdlib dependency required for advanced context encoding/decoding
Pillow>=9.0.0
